models

+++ data/models.py
@@ -21,7 +21,7 @@ class Group(Base):
     __tablename__ = 'group'
 
     id = Column(Integer, primary_key=True)
-    group_id = Column(BigInteger, nullable=False, unique=True)
+    group_id = Column(Integer, nullable=False, unique=True)
     name = Column(String, nullable=False)
     teacher_id = Column(BigInteger, ForeignKey('teacher.id'), nullable=False)
